+++ actionview/lib/action_view/template/renderable.rb
@@ -14,6 +14,8 @@ module ActionView
 
       def render(context, *args)
         @renderable.render_in(context)
+      rescue NoMethodError
+        raise ArgumentError, "'#{@renderable.inspect}' is not a renderable object. It must implement #render_in."
       end
 
       def format